separate RFQM dependencies from other services
